Dj_Project/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==4.2.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pydantic==1.7.3
pydev==0.0.1
PyJWT==2.9.0
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
